Extended grammar and generators
The grammar now distinguishes between "year" and "date". "PrintType" was added as a common type of Book and JournalArticle. That makes it simpler to handle their common attributes in the generation process. The generations were updated to work with the new grammar. Helper methods in the generators now are private